+++ src/IR/Action.hs
@@ -46,6 +46,8 @@ instance FromJSON Action where
                 case t of
                     "service_disable" -> ServiceAction <$> parseJSON v
                     "service_enable" -> ServiceAction <$> parseJSON v
+                    "service_enable_priority" -> ServiceAction <$> parseJSON v
+                    "service_disable_priority" -> ServiceAction <$> parseJSON v
                     "power_profile" -> PowerAction <$> parseJSON v
                     _unknownType -> fail $ "unknown action type: " <> T.unpack t
             )